Artikel, Kesenian, Galery policy

+++ app/Http/Controllers/KesenianController.php
@@ -121,6 +121,8 @@ class KesenianController extends Controller
      */
     public function destroy(Kesenian $kesenian)
     {
+        $this->authorize('delete', $artikel);
+        
         if ($kesenian->banner_image) {
             Storage::disk('public')->delete($kesenian->banner_image);
         }
